Results of Python Meeting on 5/18
DVD2006-2018woSuppressedwCallNumNoDiac.txt is the result of two combined searches of DVD call numbers in VgerSele (leaving out suppressed records, and adding in all of the call numbers in the associated holdings records). It contains 9649 records, and we have 11301 DVD records according to Voyager, so this should be, essentially, a complete list of our current collection.

It was processed in OpenRefine. That project was exported and can be viewed in DVD2006-2018woSuppressedwCallNumNoDiac-txt.openrefine.tar.gz.

FemaleDirectors_rev3.xlsx is a list of the female directors found on a handful of different sites. It is not a complete list, and part of our project should be to add to it. (see step-by-step for details)

FemaleDirectors_3namesource_245 is how we compared these lists. It is an unsophisticated technique, and a script may be a better options eventually. We compared the list of the names and number of occurrences pulled from the text facet of the column named "DrtsNames245" in OpenRefine in DVD2006-2018woSuppressedwCallNumNoDiac-txt.openrefine.tar.gz to the list of female director's names. We then compared this list with the number of occurrences.

To do the final processing, I had to save only the compare tab as a txt file, because Excel was crashing. FemaleDirectors_justCompare allowed me to sort the final resulting data, and I also tabulated the results.

Findings of this attempt are also included in VoyagerItemCounts.